+++ android/src/main/kotlin/com/os/operando/advertisingid/AdvertisingIdPlugin.kt
@@ -2,19 +2,19 @@ package com.os.operando.advertisingid
 
 import android.app.Activity
 import com.google.android.gms.ads.identifier.AdvertisingIdClient
-import io.flutter.embedding.engine.plugins.activity.ActivityAware
-import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.FlutterPlugin.FlutterPluginBinding
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.PluginRegistry.Registrar
 import kotlin.concurrent.thread
 
 class AdvertisingIdPlugin() : FlutterPlugin, ActivityAware, MethodCallHandler {
-    private lateinit var activity: Activity
+    private var activity: Activity? = null
 
     companion object {
         @JvmStatic
@@ -48,31 +48,33 @@ class AdvertisingIdPlugin() : FlutterPlugin, ActivityAware, MethodCallHandler {
     }
 
     override fun onMethodCall(call: MethodCall, result: Result) {
-        if (!::activity.isInitialized) {
-            result.error("noActivity", "Activity not initialized", null)
-            return 
+        if (activity == null) {
+            result.error("noActivity", "Activity is null", null)
+            return
         }
+        val a = activity!!
         when (call.method) {
             "getAdvertisingId" -> thread {
                 try {
-                    val id = AdvertisingIdClient.getAdvertisingIdInfo(activity).id
-                    activity.runOnUiThread {
+                    val id = AdvertisingIdClient.getAdvertisingIdInfo(a).id
+                    a.runOnUiThread {
                         result.success(id)
                     }
                 } catch (e: Exception) {
-                    activity.runOnUiThread {
+                    a.runOnUiThread {
                         result.error(e.javaClass.canonicalName, e.localizedMessage, null)
                     }
                 }
             }
             "isLimitAdTrackingEnabled" -> thread {
                 try {
-                    val isLimitAdTrackingEnabled = AdvertisingIdClient.getAdvertisingIdInfo(activity).isLimitAdTrackingEnabled
-                    activity.runOnUiThread {
+                    val isLimitAdTrackingEnabled =
+                        AdvertisingIdClient.getAdvertisingIdInfo(a).isLimitAdTrackingEnabled
+                    a.runOnUiThread {
                         result.success(isLimitAdTrackingEnabled)
                     }
                 } catch (e: Exception) {
-                    activity.runOnUiThread {
+                    a.runOnUiThread {
                         result.error(e.javaClass.canonicalName, e.localizedMessage, null)
                     }
                 }