+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [2.4.0] 2023-01-26
+
+* Update versions
+  * kotlin
+  * Android Gradle Plugin
+  * compileSdkVersion
+  * play-services-analytics
+
+## [2.3.0] 2022-05-24
+
+* [Fix type mismatch](https://github.com/operando/advertising_id/pull/30)
+* [Update Android dependency lib version](https://github.com/operando/advertising_id/pull/31)
+  * kotlin
+  * Android Gradle Plugin
+
 ## [2.2.0] 2022-02-11
 
 * [Update Android dependency lib version](https://github.com/operando/advertising_id/pull/26)